// src/interfaces/MenusInterfaces.tsx
// Generated by https://quicktype.io

export interface MenuDB {
    banner:     string[];
    categories: Category[];
    food:       Food[];
}

export interface Category {
    id:    number;
    name:  string;
    color: string;
    image: string;
}

export interface Food {
    name:      string;
    price:     number;
    categorie: number;
    image:     string;
}